+++ NEWS.md
@@ -9,6 +9,10 @@
 * `order_by()` gives an informative hint if you accidentally call it instead
   of `arrange()` #3357.
 
+* `filter()` handles data frame results when all columns are logical vectors
+  by reducing them with `&` (#4678). In particular this means `across()` can be used
+  in `filter()`. `filter()` gains better error messages with more information. 
+
 * `ungroup()` can now selectively remove grouping variables (#3760).
 
 * `pull()` can now return named vectors by specifying an additional column name